ios: centralize MockEnvoyEngine
We don't need to replicate this type in multiple places.
