+++ library/swift/test/EnvoyClientBuilderTests.swift
@@ -18,25 +18,6 @@ mock_template:
   virtual_clusters: {{ virtual_clusters }}
 """
 
-private final class MockEnvoyEngine: NSObject, EnvoyEngine {
-  static var onRunWithConfig: ((_ config: EnvoyConfiguration, _ logLevel: String?) -> Void)?
-  static var onRunWithYAML: ((_ configYAML: String, _ logLevel: String?) -> Void)?
-
-  func run(withConfig config: EnvoyConfiguration, logLevel: String) -> Int32 {
-    MockEnvoyEngine.onRunWithConfig?(config, logLevel)
-    return 0
-  }
-
-  func run(withConfigYAML configYAML: String, logLevel: String) -> Int32 {
-    MockEnvoyEngine.onRunWithYAML?(configYAML, logLevel)
-    return 0
-  }
-
-  func startStream(with callbacks: EnvoyHTTPCallbacks) -> EnvoyHTTPStream {
-    return MockEnvoyHTTPStream(handle: 0, callbacks: callbacks)
-  }
-}
-
 final class EnvoyClientBuilderTests: XCTestCase {
   override func tearDown() {
     super.tearDown()
